Add JdbiAccess class which allows tests to bypass packaging restrictions
