+++ core/src/main/java/org/jdbi/v3/Handle.java
@@ -27,6 +27,7 @@ import org.jdbi.v3.extension.ExtensionConfig;
 import org.jdbi.v3.extension.ExtensionFactory;
 import org.jdbi.v3.extension.NoSuchExtensionException;
 import org.jdbi.v3.mapper.ColumnMapperFactory;
+import org.jdbi.v3.mapper.DefaultMapper;
 import org.jdbi.v3.mapper.RowMapperFactory;
 import org.jdbi.v3.rewriter.StatementRewriter;
 import org.jdbi.v3.tweak.ArgumentFactory;
