+++ core/src/test/java/org/jdbi/v3/mapper/FieldMapperTest.java
@@ -11,8 +11,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-
-package org.jdbi.v3;
+package org.jdbi.v3.mapper;
 
 import static org.junit.Assert.assertEquals;
 import static org.junit.Assert.assertNotNull;
@@ -24,7 +23,13 @@ import java.sql.ResultSet;
 import java.sql.ResultSetMetaData;
 import java.sql.SQLException;
 
-import org.jdbi.v3.mapper.BeanMapper;
+import org.jdbi.v3.DerivedBean;
+import org.jdbi.v3.Handle;
+import org.jdbi.v3.JdbiAccess;
+import org.jdbi.v3.SampleBean;
+import org.jdbi.v3.StatementContext;
+import org.jdbi.v3.ValueType;
+import org.jdbi.v3.ValueTypeMapper;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
@@ -32,8 +37,9 @@ import org.mockito.Mock;
 import org.mockito.junit.MockitoJUnit;
 import org.mockito.junit.MockitoRule;
 
-public class BeanMapperTest {
-    
+
+public class FieldMapperTest {
+
     @Rule
     public MockitoRule mockitoRule = MockitoJUnit.rule();
 
@@ -43,48 +49,20 @@ public class BeanMapperTest {
     @Mock
     ResultSetMetaData resultSetMetaData;
 
-    JdbiConfig config = new JdbiConfig();
-    StatementContext ctx = new StatementContext(config);
+    Handle handle = JdbiAccess.createHandle();
+    StatementContext ctx = JdbiAccess.createContext(handle);
 
-    BeanMapper<SampleBean> mapper = new BeanMapper<>(SampleBean.class);
-    
+    FieldMapper<SampleBean> mapper = new FieldMapper<>(SampleBean.class);
+
     @Before
     public void setUpMocks() throws SQLException {
         when(resultSet.getMetaData()).thenReturn(resultSetMetaData);
     }
 
     @Test
-    public void shouldSetValueOnPublicSetter() throws Exception {
+    public void shouldSetValueOnPrivateField() throws Exception {
         when(resultSetMetaData.getColumnCount()).thenReturn(1);
         when(resultSetMetaData.getColumnLabel(1)).thenReturn("longField");
-
-        Long aLongVal = 100L;
-        when(resultSet.getLong(1)).thenReturn(aLongVal);
-        when(resultSet.wasNull()).thenReturn(false);
-
-        SampleBean sampleBean = mapper.map(resultSet, ctx);
-
-        assertEquals(aLongVal, sampleBean.getLongField());
-    }
-
-    @Test
-    public void shouldHandleColumNameWithUnderscores() throws Exception {
-        when(resultSetMetaData.getColumnCount()).thenReturn(1);
-        when(resultSetMetaData.getColumnLabel(1)).thenReturn("LONG_FIELD");
-
-        Long aLongVal = 100L;
-        when(resultSet.getLong(1)).thenReturn(aLongVal);
-        when(resultSet.wasNull()).thenReturn(false);
-
-        SampleBean sampleBean = mapper.map(resultSet, ctx);
-
-        assertEquals(aLongVal, sampleBean.getLongField());
-    }
-
-    @Test
-    public void shouldBeCaseInSensitiveOfColumnWithUnderscoresAndPropertyNames() throws Exception {
-        when(resultSetMetaData.getColumnCount()).thenReturn(1);
-        when(resultSetMetaData.getColumnLabel(1)).thenReturn("LoNg_FiElD");
 
         Long aLongVal = 100L;
         when(resultSet.getLong(1)).thenReturn(aLongVal);
@@ -105,13 +83,14 @@ public class BeanMapperTest {
     }
 
     @Test
-    public void shouldBeCaseInSensitiveOfColumnAndPropertyNames() throws Exception {
+    public void shouldBeCaseInSensitiveOfColumnAndFieldNames() throws Exception {
         when(resultSetMetaData.getColumnCount()).thenReturn(1);
         when(resultSetMetaData.getColumnLabel(1)).thenReturn("LoNgfielD");
 
         Long aLongVal = 100L;
         when(resultSet.getLong(1)).thenReturn(aLongVal);
         when(resultSet.wasNull()).thenReturn(false);
+
         SampleBean sampleBean = mapper.map(resultSet, ctx);
 
         assertEquals(aLongVal, sampleBean.getLongField());
@@ -133,57 +112,33 @@ public class BeanMapperTest {
     }
 
     @Test
-    public void shouldSetValuesOnPublicSetter() throws Exception {
-        when(resultSetMetaData.getColumnCount()).thenReturn(1);
+    public void shouldSetValuesOnAllFieldAccessTypes() throws Exception {
+        when(resultSetMetaData.getColumnCount()).thenReturn(4);
         when(resultSetMetaData.getColumnLabel(1)).thenReturn("longField");
+        when(resultSetMetaData.getColumnLabel(2)).thenReturn("protectedStringField");
+        when(resultSetMetaData.getColumnLabel(3)).thenReturn("packagePrivateIntField");
+        when(resultSetMetaData.getColumnLabel(4)).thenReturn("privateBigDecimalField");
 
-        Long expected = 1L;
-        when(resultSet.getLong(1)).thenReturn(expected);
+        Long aLongVal = 100L;
+        String aStringVal = "something";
+        int aIntVal = 1;
+        BigDecimal aBigDecimal = BigDecimal.TEN;
+        when(resultSet.getLong(1)).thenReturn(aLongVal);
+        when(resultSet.getString(2)).thenReturn(aStringVal);
+        when(resultSet.getInt(3)).thenReturn(aIntVal);
+        when(resultSet.getBigDecimal(4)).thenReturn(aBigDecimal);
         when(resultSet.wasNull()).thenReturn(false);
 
         SampleBean sampleBean = mapper.map(resultSet, ctx);
 
-        assertEquals(expected, sampleBean.getLongField());
-    }
-
-    @Test(expected = IllegalArgumentException.class)
-    public void shouldThrowOnProtectedSetter() throws Exception {
-        when(resultSetMetaData.getColumnCount()).thenReturn(1);
-        when(resultSetMetaData.getColumnLabel(1)).thenReturn("protectedStringField");
-
-        String expected = "string";
-        when(resultSet.getString(1)).thenReturn(expected);
-        when(resultSet.wasNull()).thenReturn(false);
-
-        mapper.map(resultSet, ctx);
-    }
-
-    @Test(expected = IllegalArgumentException.class)
-    public void shouldThrowOnPackagePrivateSetter() throws Exception {
-        when(resultSetMetaData.getColumnCount()).thenReturn(1);
-        when(resultSetMetaData.getColumnLabel(1)).thenReturn("packagePrivateIntField");
-
-        int expected = 200;
-        when(resultSet.getInt(1)).thenReturn(expected);
-        when(resultSet.wasNull()).thenReturn(false);
-
-        mapper.map(resultSet, ctx);
-    }
-
-    @Test(expected = IllegalArgumentException.class)
-    public void shouldThrowOnPrivateSetter() throws Exception {
-        when(resultSetMetaData.getColumnCount()).thenReturn(1);
-        when(resultSetMetaData.getColumnLabel(1)).thenReturn("privateBigDecimalField");
-
-        BigDecimal expected = BigDecimal.ONE;
-        when(resultSet.getBigDecimal(1)).thenReturn(expected);
-        when(resultSet.wasNull()).thenReturn(false);
-
-        mapper.map(resultSet, ctx);
+        assertEquals(aLongVal, sampleBean.getLongField());
+        assertEquals(aBigDecimal, sampleBean.getPrivateBigDecimalField());
+        assertEquals(aIntVal, sampleBean.getPackagePrivateIntField());
+        assertEquals(aStringVal, sampleBean.getProtectedStringField());
     }
 
     @Test
-    public void shouldSetValuesInSuperClassProperties() throws Exception {
+    public void shouldSetValuesInSuperClassFields() throws Exception {
         when(resultSetMetaData.getColumnCount()).thenReturn(2);
         when(resultSetMetaData.getColumnLabel(1)).thenReturn("longField");
         when(resultSetMetaData.getColumnLabel(2)).thenReturn("blongField");
@@ -195,7 +150,7 @@ public class BeanMapperTest {
         when(resultSet.getLong(2)).thenReturn(bLongVal);
         when(resultSet.wasNull()).thenReturn(false);
 
-        BeanMapper<DerivedBean> mapper = new BeanMapper<>(DerivedBean.class);
+        FieldMapper<DerivedBean> mapper = new FieldMapper<>(DerivedBean.class);
 
         DerivedBean derivedBean = mapper.map(resultSet, ctx);
 
@@ -205,7 +160,7 @@ public class BeanMapperTest {
 
     @Test
     public void shouldUseRegisteredMapperForUnknownPropertyType() throws Exception {
-        config.mappingRegistry.addColumnMapper(new ValueTypeMapper());
+        handle.registerColumnMapper(new ValueTypeMapper());
 
         when(resultSetMetaData.getColumnCount()).thenReturn(2);
         when(resultSetMetaData.getColumnLabel(1)).thenReturn("longField");
